+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+* Yeoman-generator 1.0 compatibility
+
 ### 1.0.0 (2016-01-07)
 * initial release